Rename to ".env" (without quotes)
This file is needed for recipe search!!
Put this file, ".env", in the main directory of DeansFoodList.
It contains the key for searching for recipes. The key references an API specifically for recipe search.

When you add NUTRITIONAL INFO, you probably will need to add a key for that API here. Get a key from the EDAMAM site.